made sure augmented mask is returned as long tensor

+++ models/processing_blocks.py
@@ -191,7 +191,7 @@ class DataAugmentor(nn.Module):
 
         concatenated_batch= self.geometric_transforms(concatenated_batch)
 
-        masks = concatenated_batch[:,3,:,:]
+        masks = concatenated_batch[:,3,:,:].long()
         images = concatenated_batch[:,0:3,:,:]
 
         images = self.colour_transforms(images)
